// web/src/types/generated/bindings/AdaptiveEnvelope.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AdaptiveTaskOperation } from "./AdaptiveTaskOperation";

export type AdaptiveEnvelope = { allowed_task_operations: Array<AdaptiveTaskOperation>, fixed_outcomes: Array<string>, fixed_acceptance: Array<string>, fixed_risk_classes: Array<string>, forbidden_side_effects: Array<string>, elevated_operations: Array<string>, };